Brush Assets: Versioning to activate new brush tool and asset
Activate the new brush tool when opening old files, and activate the
default brush asset for the mode. I'm guessing we may end up storing
the default brush asset for each mode somewhere else in the long term,
but doing it in versioning is simple now, and has the benefit of decoupling
the one time change from the defaults in the future.

Also replace the workspace's tool reference that could still point to the
old separate tools for each brush. This makes everything work by default
when opening a new factory startup file.