GP3 Undo: Add naive prototype to detect when a GeometryDrawing is actually modified.
The idea is to avoid having to trash the geometry cache when undoing a
drawing which geometry did not actually change.

This commit relies essentially on some key numbers (number of curves,
number of points, number of data layers), and comparison of implicitely
shared pointers for the data themselves.

Since the shared data needs to be 'unshared' when written to (i.e.
edited), the assumption is that different pointers mean modified data.

-----------------

NOTE: This commit is mainly intended as a quick prototype, to see if
this can be a good idea and it can fits current code.

It seems to work fine for 'selection' undos, but oddly enough any
editing to one point of one drawing will detect all others as modified
as well (apparently on detected unmatching points data).