WIP: Refactor of 'partial blendfile write' BKE code.
This commit introduces a new 'PartialWriteContext' class, which wraps
around a regular Main struct. It is designed to make writing a set of
IDs easy and safe, and to prepare for future 'asset library editing'
low-level code.

The main goal of this refactor is to provide the same functionalities
(or better ones) than existing partial write code, without the very bad
hacks currently done.

Essentially, it allows to:
  * Add (aka copy) IDs from the G_MAIN to the partial write context.
    * This process handles dependencies and libraries automatically.
    * A refined handling of dependencies is possible through an
      optional 'filtering' callback.
  * Keep track of added IDs, to allow de-duplication in case data is
    added more than once.
  * Cleanup the context (i.e. remove unused IDs).
  * Write the context to disk as a blendfile.

Since the context keeps information to find matches between its content
and IDs from the G_MAIN, its lifespan is expected to be _very_ short.
Otherwise, changes in G_MAIN (relationships between IDs, their session
uuid, etc.) cannot be tracked by the context, leading to inconsistencies.
A partial write context should typically be created, filled, wrtitten
and deleted within a same function.

NOTE: As a demo, the copy buffer for the VSE and the python's API to
write into libraries have been update to use the new partial write
context.