UI: Generic String Wrapping
Given a string, font id, and width, return a string vector with the string
wrapped to that width.

---

We currently only have string wrapping that occurs at the time of text output. But it can be useful to know where a string would break without printing it. For wrapping of uiBut labels, multi-line text input, etc.