VSE: Speedup isffmpeg by doing a simpler check
isffmpeg is being used to answer the question, "is this a video file?".
It is generally called already on files that are known to have video
extensions (e.g. .mp4), and was doing a full-blown "initialize
libavformat decoding context, read header, decode some frames just to
be sure, find video stream, see if we have a suitable video stream
decoder".

Now changed that to only do the "initialize decoding context, read
header" and if that is successful, assume it is a file that ffmpeg
can understand. In theory this can change output from "false"
previously to "true" now, for files that are corrupted or have no video
streams. In practice, all that is moot since the later "try to playback"
code will do all that checking anyway.

In tests of VSE with broken / no-video files, end-user result looks
exactly the same, e.g. dragging a no-video file into VSE says
"not an anim" just like before.

In Gold previs playback between 1:41-1:55, on Windows/Ryzen5950X:
- Slowest 3 frames went from 276, 195, 168ms -> 239, 173, 147ms
  (saves ~20ms per frame). All of these frames are camera cuts where
  multiple new video clips start playing.
- In the whole playback, total amount of time taken by
  `imb_get_anim_type`: 234ms -> 42ms.