EEVEE-Next: Update reflection probes when resolution changes
When the resolution of the reflection probes changes the change
was only applied when the reflection probes where updated. To improve
the workflow this change will update the reflection probes when the
resolution is different.

Changing resolution would change the location it has on the atlas and
they are temporary not used during rendering. This can lead to
flickering at the first sample.

NOTE: the flickering can be reduced by still locking the previous
location on the atlas, but that has some draw backs when working
with scenes having many reflection probes.