Compositor: Unify Keying node between CPU and GPU
This patch unifies the Keying node between the CPU and GPU. The
difference was due to imprecision when computing pixel saturation in the
CPU code. This typically happens when the pixel is gray scale, and
should have a zero saturation, but it came out negative due to
imprecision. To fix this, evaluate the mix expression using doubles.