Fix #116598: Viewer node crash
The crash occurs when a user tries to connect
the viewer node to a curve attribute while the
viewer is set to the node type 'Quaternion'.

The problem stems from a failed attribute lookup
due to a failed conversion from 'ColorGeometry4f'
to 'Quaternion', returning NULL. There was no
NULL check after the lookup, so it was resulting
in a crash in this specific instance.

I fixed the problem by adding an empty parameter
check to the lookup, since it should always be
returning the desired attribute.