Fix #103824: CompositorNodeOutputFile inputs.new() crash
Even though this has a more dedicated API in `file_slots.new()` this
shouldnt crash.

The crash is due to `NodeImageMultiFileSocket` not being initialized
properly, `ntreeBlendWrite` then crashes accessing garbage data.

Note this is more of a bandaid, the proper way forward would be some
sort of dynamic declration such as
`blender::nodes::node_group_declare_dynamic` as outlined in #108728, but
since this is a more involved change, this PR just fixes the crash with
what we have atm. (using `ntreeCompositOutputFileAddSocket` which takes
care of `NodeImageMultiFileSocket` data for us).