EEVEE-Next: Refresh when resizing area light via gizmo
When using the area light gizmo in the 3d viewport to resize the
area light, the changes where only made visible when EEVEE-Next would
rerender. EEVEE and Cycles did refresh accordingly.

The reason is that EEVEE and Cycles uses a more general approach of
detecting changes. EEVEE-Next listens to exact changes made in the
dependency graph.

This change fixes this by tagging the lamp without any flags similar
to what RNA does. EEVEE and Cycles don't need the general flag for
correct working.