Fix: Remove duplicate image_add operators in the 3D view port
As part of #118623, we discovered that the operator used for adding
images in the 3D viewport was not the same when drag and dropping as
when adding it from the add menu.
This lead to different and potentially confusing behaviour for the user
when they use both.

This patch removes the python operator and unifies the functionality
into one operator. It also renames the operator to be in line with the
other "Add X" object operators.