Fix: Tree-view items always collapsed by default
Since 660867fa00, having tree-view items uncollapsed by default using
the `set_collapsed()` functions wouldn't work anymore. An attempt to do
this would assert even, so eb71d9f7bc disabled the assert.

I think a function designed to handle exactly this is the best solution,
it makes the intent & behavior more clear than before, and highlights
that this is a special case.

It's a bit of a shame that something as simple as the collapsed state is
becoming so tricky to manage. I have some ideas to improve this, but
that will require some more work.

Fixes #117957.